Use float instead of double in WardDist::pdf
git-archimport-id: miles@gnu.org--2006/snogray--devo--0--patch-298

+++ ward-dist.h
@@ -51,10 +51,10 @@ public:
       return 0;
     else
       {
-	double inv_cos_theta = 1 / cos_theta;
-	double inv_cos2_theta = inv_cos_theta * inv_cos_theta;
-	double inv_cos3_theta = inv_cos2_theta * inv_cos_theta;
-	double tan2_theta = inv_cos2_theta - 1;
+	float inv_cos_theta = 1 / cos_theta;
+	float inv_cos2_theta = inv_cos_theta * inv_cos_theta;
+	float inv_cos3_theta = inv_cos2_theta * inv_cos_theta;
+	float tan2_theta = inv_cos2_theta - 1;
 	return inv_m2 * M_1_PIf * inv_cos3_theta * exp (-tan2_theta * inv_m2);
       }
   }
